+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/TextArea/tests/TextArea.test.js
@@ -1,6 +1,6 @@
 // @flow
 import React from 'react';
-import {render} from '@testing-library/react';
+import {fireEvent, render, screen} from '@testing-library/react';
 import TextArea from '../TextArea';
 
 jest.mock('../../../utils/Translator', () => ({
@@ -42,24 +42,32 @@ test('TextArea should render with value and character counter', () => {
     expect(container).toMatchSnapshot();
 });
 
-// test('TextArea should call onBlur when it loses focus', () => {
-//     const blurSpy = jest.fn();
-//     const textArea = shallow(<TextArea onBlur={blurSpy} onChange={jest.fn()} value="" />);
+test('TextArea should call onBlur when it loses focus', () => {
+    const blurSpy = jest.fn();
+    render(<TextArea onBlur={blurSpy} onChange={jest.fn()} value="" />);
 
-//     textArea.find('textarea').simulate('blur');
-//     expect(blurSpy).toBeCalledWith();
-// });
+    const textarea = screen.queryByRole('textbox');
+    fireEvent.blur(textarea);
 
-// test('TextArea should call onChange when the TextArea changes', () => {
-//     const changeSpy = jest.fn();
-//     const textArea = shallow(<TextArea onChange={changeSpy} value="My value" />);
-//     textArea.find('textarea').simulate('change', {currentTarget: {value: 'my-value'}});
-//     expect(changeSpy).toHaveBeenCalledWith('my-value');
-// });
+    expect(blurSpy).toBeCalledWith();
+});
 
-// test('TextArea should call onChange with undefined when the TextArea changes to empty', () => {
-//     const changeSpy = jest.fn();
-//     const textArea = shallow(<TextArea onChange={changeSpy} value="My value" />);
-//     textArea.find('textarea').simulate('change', {currentTarget: {value: ''}});
-//     expect(changeSpy).toHaveBeenCalledWith(undefined);
-// });
+test('TextArea should call onChange when the TextArea changes', () => {
+    const changeSpy = jest.fn();
+    render(<TextArea onChange={changeSpy} value="My value" />);
+
+    const textarea = screen.queryByDisplayValue('My value');
+    fireEvent.change(textarea, {target: {value: 'my-value'}});
+
+    expect(changeSpy).toHaveBeenCalledWith('my-value');
+});
+
+test('TextArea should call onChange with undefined when the TextArea changes to empty', () => {
+    const changeSpy = jest.fn();
+    render(<TextArea onChange={changeSpy} value="My value" />);
+
+    const textarea = screen.queryByDisplayValue('My value');
+    fireEvent.change(textarea, {target: {value: ''}});
+
+    expect(changeSpy).toHaveBeenCalledWith(undefined);
+});
